Move path finder to separate templating module
